--- src/components/A11y/A11y.ts
@@ -1,14 +1,0 @@
-import styled from 'styled-components';
-
-const A11y = styled.h1`
-  border: 0;
-  clip: rect(0 0 0 0);
-  height: 1px;
-  margin: -1px;
-  overflow: hidden;
-  padding: 0;
-  position: absolute;
-  width: 1px;
-`;
-
-export default A11y;
