--- src/components/A11y/index.ts
@@ -1,1 +1,0 @@
-export { default } from './A11y';
