--- src/containers/Home/components/Banner/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Banner';
