+++ src/styles/globals.css
@@ -14,12 +14,12 @@
   --gray-light: #f5f5f5;
   --julge-black: #111322;
 }
-
+/* 
 .text-gradient-bg {
   -webkit-background-clip: text;
   background-clip: text;
   -webkit-text-fill-color: transparent;
-}
+} */
 
 /* 
 @media (prefers-color-scheme: dark) {
